Refactor user-related services and controllers to improve organization
