+++ server/src/main/java/com/ultracards/server/service/auth/AuthService.java
@@ -1,18 +1,15 @@
 package com.ultracards.server.service.auth;
 
-import com.ultracards.gateway.dto.EmailDTO;
 import com.ultracards.gateway.dto.auth.GameStatsDTO;
 import com.ultracards.gateway.dto.auth.ProfileDTO;
-import com.ultracards.gateway.dto.auth.UserSessionDTO;
 import com.ultracards.gateway.dto.auth.UserGamesStatsDTO;
 import com.ultracards.gateway.dto.auth.UsernameDTO;
 import com.ultracards.gateway.dto.auth.VerificationCodeDTO;
 import com.ultracards.server.entity.UserEntity;
-import com.ultracards.server.entity.auth.UserSession;
 import com.ultracards.server.enums.games.GameType;
 import com.ultracards.server.repositories.UserRepository;
 import com.ultracards.server.service.EmailService;
-import com.ultracards.server.service.UserService;
+import com.ultracards.server.service.users.UserService;
 import com.ultracards.server.service.games.UserGamesStatsService;
 import jakarta.mail.MessagingException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -27,7 +24,6 @@ import org.springframework.stereotype.Service;
 import java.io.UnsupportedEncodingException;
 import java.time.Instant;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -38,9 +34,7 @@ public class AuthService {
     private final VerificationCodeService verificationCodeService;
     private final UserService userService;
     private final UserRepository userRepository;
-    private final TokenService tokenService;
     private final EmailService emailService;
-    private final UserGamesStatsService userGamesStatsService;
     private final SessionService sessionService;
 
     @Value("${app.cookie-token.same-site:Lax}")
@@ -124,71 +118,6 @@ public class AuthService {
         }
 
         return res;
-    }
-
-    public ProfileDTO getProfile(UserEntity user) {
-        return createProfileByUser(user);
-    }
-
-    public Boolean updateProfile(
-            UserEntity user,
-            @Valid ProfileDTO profileDTO,
-            String token) {
-        return updateProfileByUser(user, profileDTO, token);
-    }
-
-    private Boolean updateProfileByUser(UserEntity user, @Valid ProfileDTO profileDTO, String token) {
-        var session = sessionService.getSession(token);
-        var authenticatedAt = session.getLastAuthenticatedAt();
-        if (Instant.now().isAfter(authenticatedAt.plusSeconds(updateDuration * 60)))
-            return false;
-        user.setUsername(profileDTO.getUsername());
-        user.setEmail(profileDTO.getEmail());
-        userRepository.save(user);
-        return true;
-    }
-
-    private ProfileDTO createProfileByUser(UserEntity user) {
-        var profile = new ProfileDTO();
-
-        profile.setUsername(user.getUsername());
-        profile.setEmail(user.getEmail());
-        profile.setRoles(user.getRoles().stream().map(Enum::toString).collect(Collectors.toList()));
-        profile.setId(user.getId());
-
-        var gameStats = userGamesStatsService.getByUser(user);
-
-        if (gameStats == null) {
-            return profile;
-        }
-        var games = new LinkedHashMap<String, GameStatsDTO>();
-        for (var gameType : GameType.values()) {
-            games.put(
-                    gameType.name(),
-                    new GameStatsDTO(
-                            gameStats.getGamesPlayed(gameType),
-                            gameStats.getGamesWon(gameType),
-                            gameStats.getLastPlayedAt(gameType)
-                    )
-            );
-        }
-
-        profile.setUserGamesStats(new UserGamesStatsDTO(
-                gameStats.getId(),
-                user.getId(),
-                games
-        ));
-
-        var gamesPlayed = 0;
-        var gamesWon = 0;
-        for (var stats: games.values()) {
-            gamesPlayed += stats.getPlayed();
-            gamesWon += stats.getWins();
-        }
-        profile.setGamesPlayed(gamesPlayed);
-        profile.setGamesWon(gamesWon);
-
-        return profile;
     }
 
     private void expireRefreshTokenCookie(HttpServletResponse response, String path) {
